Ensure Pac4JUserProfiles are ClusterSerializable
